event_manager/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==25.1.0
click==7.1.2
colorama==0.4.6
colorclass==2.2.2
colorlog==6.9.0
croniter==6.0.0
Django==4.2.19
django-filter==24.3
djangorestframework==3.15.2
faust==1.10.4
frozenlist==1.5.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
kafka-python==1.4.7
mode==4.3.2
multidict==6.1.0
mypy-extensions==1.0.0
opentracing==1.3.0
promise==2.3
propcache==0.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
robinhood-aiokafka==1.1.6
six==1.17.0
sqlparse==0.5.3
terminaltables==3.1.10
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
venusian==1.2.0
yarl==1.15.2